pyproject.toml
[project]
name = "silico"
version = "0.1.0"
description = "Society in Silico - book tools and infrastructure"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
silico-tts = "silico.tts_pipeline:main"
silico-tts-enhanced = "silico.tts_enhanced:main"
silico-check = "silico.concept_tracker:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/silico"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["src"]
